+++ src/carve/BuiltinCarvers.cpp
@@ -6,6 +6,7 @@
 #include "formats/JpegCarver.hpp"
 #include "formats/Mp4Carver.hpp"
 #include "formats/PngCarver.hpp"
+#include "formats/RawCarver.hpp"
 #include "revenant/carve/CarverRegistry.hpp"
 
 namespace revenant::carve {
@@ -14,6 +15,7 @@ void registerBuiltinCarvers(CarverRegistry& registry) {
 	registry.registerCarver(std::make_unique<JpegCarver>());
 	registry.registerCarver(std::make_unique<PngCarver>());
 	registry.registerCarver(std::make_unique<Mp4Carver>());
+	registry.registerCarver(std::make_unique<RawCarver>());
 }
 
 } // namespace revenant::carve